+++ client/web-sveltekit/src/routes/[...repo=reporev]/(validrev)/(code)/-/tree/[...path]/+page.ts
@@ -1,13 +1,10 @@
-import { fetchBlobPlaintext } from '$lib/repo/api/blob'
-import { fetchDiff } from '$lib/repo/api/commits'
 import { fetchTreeEntries } from '$lib/repo/api/tree'
 import { findReadme } from '$lib/repo/tree'
 
 import type { PageLoad } from './$types'
-import { TreeEntriesCommitInfo } from './page.gql'
+import { TreePageCommitInfoQuery, TreePageReadmeQuery } from './page.gql'
 
-export const load: PageLoad = async ({ params, parent, url }) => {
-    const revisionToCompare = url.searchParams.get('rev')
+export const load: PageLoad = async ({ params, parent }) => {
     const { resolvedRevision, graphqlClient } = await parent()
 
     const treeEntries = fetchTreeEntries({
@@ -26,7 +23,7 @@ export const load: PageLoad = async ({ params, parent, url }) => {
             treeEntries,
             commitInfo: graphqlClient
                 .query({
-                    query: TreeEntriesCommitInfo,
+                    query: TreePageCommitInfoQuery,
                     variables: {
                         repoID: resolvedRevision.repo.id,
                         commitID: resolvedRevision.commitID,
@@ -48,21 +45,22 @@ export const load: PageLoad = async ({ params, parent, url }) => {
                 if (!readme) {
                     return null
                 }
-                return fetchBlobPlaintext({
-                    repoID: resolvedRevision.repo.id,
-                    commitID: resolvedRevision.commitID,
-                    filePath: readme.path,
-                }).then(result => ({
-                    name: readme.name,
-                    ...result,
-                }))
+                return graphqlClient
+                    .query({
+                        query: TreePageReadmeQuery,
+                        variables: {
+                            repoID: resolvedRevision.repo.id,
+                            revspec: resolvedRevision.commitID,
+                            path: readme.path,
+                        },
+                    })
+                    .then(result => {
+                        if (result.data.node?.__typename !== 'Repository') {
+                            throw new Error('Expected Repository')
+                        }
+                        return result.data.node.commit?.blob ?? null
+                    })
             }),
-            compare: revisionToCompare
-                ? {
-                      revisionToCompare,
-                      diff: fetchDiff(resolvedRevision.repo.id, revisionToCompare, [params.path]),
-                  }
-                : null,
         },
     }
 }