+++ cmd/frontend/internal/context/resolvers/context_test.go
@@ -33,10 +33,20 @@ import (
 func TestContextResolver(t *testing.T) {
 	logger := logtest.Scoped(t)
 	ctx := context.Background()
-
 	db := database.NewDB(logger, dbtest.NewDB(t))
 	repo1 := types.Repo{Name: "repo1"}
 	repo2 := types.Repo{Name: "repo2"}
+	truePtr := true
+	conf.Mock(&conf.Unified{
+		SiteConfiguration: schema.SiteConfiguration{
+			CodyEnabled: &truePtr,
+			Embeddings: &schema.Embeddings{
+				Provider:    "sourcegraph",
+				Enabled:     &truePtr,
+				AccessToken: "123",
+			},
+		},
+	})
 	// Create populates the IDs in the passed in types.Repo
 	err := db.Repos().Create(ctx, &repo1, &repo2)
 	require.NoError(t, err)
@@ -138,13 +148,6 @@ func TestContextResolver(t *testing.T) {
 		mockGitserver,
 		contextClient,
 	)
-
-	truePtr := true
-	conf.Mock(&conf.Unified{
-		SiteConfiguration: schema.SiteConfiguration{
-			CodyEnabled: &truePtr,
-		},
-	})
 
 	ctx = actor.WithActor(ctx, actor.FromMockUser(1))
 	ffs := featureflag.NewMemoryStore(map[string]bool{"cody": true}, nil, nil)