+++ cmd/frontend/internal/dotcom/productsubscription/licenses_db.go
@@ -99,7 +99,6 @@ func (s dbLicenses) Create(ctx context.Context, subscriptionID, licenseKey strin
 	}
 
 	if featureflag.FromContext(ctx).GetBoolOr("auditlog-expansion", false) {
-
 		// Log an event when a license is created in DotCom
 		if err := s.db.SecurityEventLogs().LogSecurityEvent(ctx, database.SecurityEventNameDotComLicenseCreated, "", uint32(actor.FromContext(ctx).UID), "", "BACKEND", nil); err != nil {
 			logger.Warn("Error logging security event", log.Error(err))
@@ -195,7 +194,7 @@ func (s dbLicenses) GetByID(ctx context.Context, id string) (*dbLicense, error) 
 	return results[0], nil
 }
 
-// GetByLicenseKey retrieves the product license (if any) given its check license token.
+// GetByAccessToken retrieves the product license (if any) given its check license token.
 // The accessToken is of the format created by GenerateLicenseKeyBasedAccessToken.
 //
 // 🚨 SECURITY: The caller must ensure that errTokenInvalid error is handled appropriately
@@ -218,7 +217,7 @@ func (s dbLicenses) GetByAccessToken(ctx context.Context, accessToken string) (*
 	return results[0], nil
 }
 
-// GetByID retrieves the product license (if any) given its license key.
+// GetByLicenseKey retrieves the product license (if any) given its license key.
 func (s dbLicenses) GetByLicenseKey(ctx context.Context, licenseKey string) (*dbLicense, error) {
 	if mocks.licenses.GetByLicenseKey != nil {
 		return mocks.licenses.GetByLicenseKey(licenseKey)
@@ -286,19 +285,26 @@ func (s dbLicenses) Active(ctx context.Context, subscriptionID string) (*dbLicen
 	return licenses[0], nil
 }
 
-// AssignSiteID marks the existing license as used by a specific siteID
-func (s dbLicenses) AssignSiteID(ctx context.Context, id, siteID string) error {
+// AssignSiteID marks the existing license as used by a specific siteID, and
+// returns the updated license. The original dbLicense struct is modified.
+func (s dbLicenses) AssignSiteID(ctx context.Context, license *dbLicense, siteID string) (*dbLicense, error) {
 	q := sqlf.Sprintf(`
 UPDATE product_licenses
 SET site_id = %s
 WHERE id = %s
 	`,
 		siteID,
-		id,
+		license.ID,
 	)
 
 	_, err := s.db.ExecContext(ctx, q.Query(sqlf.PostgresBindVar), q.Args()...)
-	return err
+	if err != nil {
+		return nil, err
+	}
+
+	license.SiteID = &siteID
+
+	return license, nil
 }
 
 // List lists all product licenses that satisfy the options.
@@ -373,8 +379,7 @@ ORDER BY created_at DESC
 	}
 
 	if featureflag.FromContext(ctx).GetBoolOr("auditlog-expansion", false) {
-
-		//Log an event when liscenses list is viewed in Dotcom
+		// Log an event when liscense list is viewed in Dotcom
 		if err := s.db.SecurityEventLogs().LogSecurityEvent(ctx, database.SecurityEventNameDotComLicenseViewed, "", uint32(actor.FromContext(ctx).UID), "", "BACKEND", q.Args()); err != nil {
 			logger.Warn("Error logging security event", log.Error(err))
 		}