+++ cmd/searcher/internal/search/search_regex.go
@@ -24,12 +24,12 @@ func regexSearchBatch(
 	zf *zipFile,
 	limit int,
 	patternMatchesContent, patternMatchesPaths bool,
-	isPatternNegated bool,
+	isCaseSensitive bool,
 	contextLines int32,
 ) ([]protocol.FileMatch, bool, error) {
 	ctx, cancel, sender := newLimitedStreamCollector(ctx, limit)
 	defer cancel()
-	err := regexSearch(ctx, m, pm, zf, patternMatchesContent, patternMatchesPaths, isPatternNegated, sender, contextLines)
+	err := regexSearch(ctx, m, pm, zf, patternMatchesContent, patternMatchesPaths, isCaseSensitive, sender, contextLines)
 	return sender.collected, sender.LimitHit(), err
 }
 
@@ -53,15 +53,15 @@ func regexSearch(
 	pm *pathMatcher,
 	zf *zipFile,
 	patternMatchesContent, patternMatchesPaths bool,
-	isPatternNegated bool,
+	isCaseSensitive bool,
 	sender matchSender,
 	contextLines int32,
 ) (err error) {
 	tr, ctx := trace.New(ctx, "regexSearch")
 	defer tr.EndWithErr(&err)
 
-	m.AddAttributes(tr)
 	tr.SetAttributes(attribute.Stringer("path", pm))
+	tr.SetAttributes(attribute.String("matcher", m.String()))
 
 	if !patternMatchesContent && !patternMatchesPaths {
 		patternMatchesContent = true
@@ -83,11 +83,11 @@ func regexSearch(
 		files = zf.Files
 	)
 
-	if m.MatchesAllContent() || (patternMatchesPaths && !patternMatchesContent) {
+	if _, ok := m.(allMatcher); ok || (patternMatchesPaths && !patternMatchesContent) {
 		// Fast path for only matching file paths (or with a nil pattern, which matches all files,
 		// so is effectively matching only on file paths).
 		for _, f := range files {
-			if match := pm.Matches(f.Name) && m.MatchesString(f.Name); match == !isPatternNegated {
+			if pm.Matches(f.Name) && m.MatchesString(f.Name) {
 				if ctx.Err() != nil {
 					return ctx.Err()
 				}
@@ -142,7 +142,7 @@ func regexSearch(
 				// data (for Preview).
 				fileBuf := zf.DataFor(f)
 				fileMatchBuf := fileBuf
-				if m.IgnoreCase() {
+				if !isCaseSensitive {
 					// If we are ignoring case, we transform the input instead of
 					// relying on the regular expression engine which can be
 					// slow. compilePattern has already lowercased the pattern. We also
@@ -155,9 +155,8 @@ func regexSearch(
 					casetransform.BytesToLowerASCII(fileMatchBuf, fileBuf)
 				}
 
-				locs := m.MatchesFile(fileMatchBuf, sender.Remaining())
+				match, locs := m.MatchesFile(fileMatchBuf, sender.Remaining())
 				fm := locsToFileMatch(fileBuf, f.Name, locs, contextLines)
-				match := len(fm.ChunkMatches) > 0
 				if !match && patternMatchesPaths {
 					// Try matching against the file path.
 					match = m.MatchesString(f.Name)
@@ -165,7 +164,7 @@ func regexSearch(
 						fm.Path = f.Name
 					}
 				}
-				if match == !isPatternNegated {
+				if match {
 					sender.Send(fm)
 				}
 			}