+++ internal/gitserver/mocks_temp.go
@@ -36,9 +36,6 @@ type MockClient struct {
 	// BatchLogFunc is an instance of a mock function object controlling the
 	// behavior of the method BatchLog.
 	BatchLogFunc *ClientBatchLogFunc
-	// BlameFileFunc is an instance of a mock function object controlling
-	// the behavior of the method BlameFile.
-	BlameFileFunc *ClientBlameFileFunc
 	// BranchesContainingFunc is an instance of a mock function object
 	// controlling the behavior of the method BranchesContaining.
 	BranchesContainingFunc *ClientBranchesContainingFunc
@@ -224,11 +221,6 @@ func NewMockClient() *MockClient {
 		},
 		BatchLogFunc: &ClientBatchLogFunc{
 			defaultHook: func(context.Context, BatchLogOptions, BatchLogCallback) (r0 error) {
-				return
-			},
-		},
-		BlameFileFunc: &ClientBlameFileFunc{
-			defaultHook: func(context.Context, api.RepoName, string, *BlameOptions) (r0 []*Hunk, r1 error) {
 				return
 			},
 		},
@@ -529,11 +521,6 @@ func NewStrictMockClient() *MockClient {
 				panic("unexpected invocation of MockClient.BatchLog")
 			},
 		},
-		BlameFileFunc: &ClientBlameFileFunc{
-			defaultHook: func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error) {
-				panic("unexpected invocation of MockClient.BlameFile")
-			},
-		},
 		BranchesContainingFunc: &ClientBranchesContainingFunc{
 			defaultHook: func(context.Context, api.RepoName, api.CommitID) ([]string, error) {
 				panic("unexpected invocation of MockClient.BranchesContaining")
@@ -822,9 +809,6 @@ func NewMockClientFrom(i Client) *MockClient {
 		},
 		BatchLogFunc: &ClientBatchLogFunc{
 			defaultHook: i.BatchLog,
-		},
-		BlameFileFunc: &ClientBlameFileFunc{
-			defaultHook: i.BlameFile,
 		},
 		BranchesContainingFunc: &ClientBranchesContainingFunc{
 			defaultHook: i.BranchesContaining,
@@ -1408,119 +1392,6 @@ func (c ClientBatchLogFuncCall) Args() []interface{} {
 // invocation.
 func (c ClientBatchLogFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
-}
-
-// ClientBlameFileFunc describes the behavior when the BlameFile method of
-// the parent MockClient instance is invoked.
-type ClientBlameFileFunc struct {
-	defaultHook func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error)
-	hooks       []func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error)
-	history     []ClientBlameFileFuncCall
-	mutex       sync.Mutex
-}
-
-// BlameFile delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockClient) BlameFile(v0 context.Context, v1 api.RepoName, v2 string, v3 *BlameOptions) ([]*Hunk, error) {
-	r0, r1 := m.BlameFileFunc.nextHook()(v0, v1, v2, v3)
-	m.BlameFileFunc.appendCall(ClientBlameFileFuncCall{v0, v1, v2, v3, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the BlameFile method of
-// the parent MockClient instance is invoked and the hook queue is empty.
-func (f *ClientBlameFileFunc) SetDefaultHook(hook func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// BlameFile method of the parent MockClient instance invokes the hook at
-// the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *ClientBlameFileFunc) PushHook(hook func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *ClientBlameFileFunc) SetDefaultReturn(r0 []*Hunk, r1 error) {
-	f.SetDefaultHook(func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *ClientBlameFileFunc) PushReturn(r0 []*Hunk, r1 error) {
-	f.PushHook(func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error) {
-		return r0, r1
-	})
-}
-
-func (f *ClientBlameFileFunc) nextHook() func(context.Context, api.RepoName, string, *BlameOptions) ([]*Hunk, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *ClientBlameFileFunc) appendCall(r0 ClientBlameFileFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of ClientBlameFileFuncCall objects describing
-// the invocations of this function.
-func (f *ClientBlameFileFunc) History() []ClientBlameFileFuncCall {
-	f.mutex.Lock()
-	history := make([]ClientBlameFileFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// ClientBlameFileFuncCall is an object that describes an invocation of
-// method BlameFile on an instance of MockClient.
-type ClientBlameFileFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 api.RepoName
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 string
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 *BlameOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 []*Hunk
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c ClientBlameFileFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c ClientBlameFileFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
 }
 
 // ClientBranchesContainingFunc describes the behavior when the